backend/requirements.txt
# ── Core framework ──────────────────────────────
fastapi>=0.110
uvicorn[standard]>=0.27
pydantic>=2.6
httpx>=0.27
orjson>=3.10
python-multipart>=0.0.9
python-dotenv>=1.0
anyio>=4.3

# ── Image + face detection stack ────────────────
opencv-python>=4.9
numpy>=1.26
pillow>=10.2

# ── Auth & security ─────────────────────────────
PyJWT>=2.9
passlib[bcrypt]>=1.7
bcrypt>=4.1
email-validator>=2.1
